moves ResourcesUtiles to internal
