+++ common/src/main/java/io/netty/util/internal/ResourcesUtil.java
@@ -12,7 +12,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package io.netty.util;
+package io.netty.util.internal;
 
 import java.io.File;
 import java.io.UnsupportedEncodingException;
